+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog follows the patterns described here: https://keepachangelog.com/e
 ### added
 - Added `head_classes`, `body_classes` & `foot_classes` optional properties to the `Hero` component. These extra classes will be added to the stanard classes assigned to each of the corresponding hero component sections.
 - Added `navburger_classes` prop to the `Navbar` component. These extra classes will be added to the standard classes for the `navbar-burger` element.
+- Added `help_has_error` prop to the `Field` component. This `bool` prop will add the `is-danger` class to the field's help text when `true`.
 
 ## 0.1.6
 ### changed